5192: Forbedre innlasting av oppgaveoversikt
Henter nå oppgaveoversikten kun når vi lander på forsiden. I tillegg har vi flytta 5-sekundsoppdateringen til journalføringssteget
