+++ src/ducks/navigering/operations.js
@@ -1,9 +1,6 @@
 import { push } from "connected-react-router";
 
-import { oppgaverOperations } from "../oppgaver";
-
 export const tilForsiden = () => async (dispatch) => {
-  dispatch(oppgaverOperations.oversikt());
   return dispatch(push("/"));
 };
