+++ src/sider/forside/komponenter/mineoppgaver/behandlingOppgaver.tsx
@@ -1,5 +1,5 @@
-import React, { useEffect } from "react";
-import { connect, ConnectedProps, useDispatch } from "react-redux";
+import React from "react";
+import { connect, ConnectedProps } from "react-redux";
 
 import { RootState } from "AppTypes";
 
@@ -12,7 +12,6 @@ import { landkoderSelectors } from "../../../../ducks/landkoder";
 import "./behandlingsoppgaver.css";
 import BehandlingOppgave from "../../../../felleskomponenter/oppgaveliste/behandlingOppgave";
 import { useFeatureToggle } from "../../../../featuretoggle";
-import { oversikt } from "../../../../ducks/oppgaver/operations";
 
 const mapStateToProps = (state: RootState) => ({
   mineSaker: oppgaverSelectors.MineSakerSelector(state),
@@ -30,15 +29,6 @@ export const BehandlingOppgaver = ({ mineSaker, landkoder }: PropsFromRedux) => 
   const folketrygdenToggle = useFeatureToggle("melosys.folketrygden.mvp");
 
   const { saksbehandling } = mineSaker;
-  const dispatch = useDispatch();
-
-  const refreshOversiktDelayMillis = 5000;
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      dispatch(oversikt());
-    }, refreshOversiktDelayMillis);
-    return () => clearTimeout(timer);
-  }, []);
 
   return (
     <div className="behandlingsOppgaver">
